+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/Launch.kt
@@ -7,7 +7,7 @@ import com.github.insanusmokrassar.AutoPostTelegramBot.base.models.FinalConfig
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.DefaultPluginManager
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.commonLogger
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.NewDefaultCoroutineScope
-import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.subscribe
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.flow.collectWithErrors
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.load
 import com.github.insanusmokrassar.TelegramBotAPI.requests.chat.get.GetChat
 import com.github.insanusmokrassar.TelegramBotAPI.types.CallbackQuery.MessageDataCallbackQuery
@@ -15,8 +15,12 @@ import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.MediaG
 import com.github.insanusmokrassar.TelegramBotAPI.types.update.CallbackQueryUpdate
 import com.github.insanusmokrassar.TelegramBotAPI.types.update.MediaGroupUpdates.MediaGroupUpdate
 import com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts.BaseMessageUpdate
+import com.github.insanusmokrassar.TelegramBotAPI.updateshandlers.FlowsUpdatesFilter
+import com.github.insanusmokrassar.TelegramBotAPI.utils.extensions.UpdateReceiver
 import kotlinx.coroutines.channels.BroadcastChannel
 import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.flow.asFlow
+import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 import org.jetbrains.exposed.sql.SchemaUtils
 import org.jetbrains.exposed.sql.transactions.transaction
@@ -39,13 +43,22 @@ const val extraLargeBroadcastCapacity = 64
 
 const val commonListenersCapacity = mediumBroadcastCapacity
 
+val flowFilter = FlowsUpdatesFilter()
+
+@Deprecated("Solved to use flows in the future. Use \"flowFilter\" instead")
 val allMessagesListener = BroadcastChannel<BaseMessageUpdate>(Channel.CONFLATED)
+@Deprecated("Solved to use flows in the future. Use \"flowFilter\" instead")
 val allCallbackQueryListener = BroadcastChannel<CallbackQueryUpdate>(Channel.CONFLATED)
+@Deprecated("Solved to use flows in the future. Use \"flowFilter\" instead")
 val allMediaGroupsListener = BroadcastChannel<MediaGroupUpdate>(Channel.CONFLATED)
 
 val messagesListener = BroadcastChannel<BaseMessageUpdate>(Channel.CONFLATED)
 val callbackQueryListener = BroadcastChannel<CallbackQueryUpdate>(Channel.CONFLATED)
 val mediaGroupsListener = BroadcastChannel<MediaGroupUpdate>(Channel.CONFLATED)
+
+val checkedMessagesFlow = messagesListener.asFlow()
+val checkedCallbacksQueriesFlow = callbackQueryListener.asFlow()
+val checkedMediaGroupsFlow = mediaGroupsListener.asFlow()
 
 fun main(args: Array<String>) {
     val config: FinalConfig = load(args[0], Config.serializer()).finalConfig
@@ -73,42 +86,53 @@ fun main(args: Array<String>) {
             config
         )
 
-        NewDefaultCoroutineScope().apply {
-            allMessagesListener.subscribe(
-                scope = this
-            ) {
+        NewDefaultCoroutineScope(8).apply {
+            val messageUpdatesCollector: UpdateReceiver<BaseMessageUpdate> = {
+                allMessagesListener.offer(it)
                 if (it.data.chat.id == config.sourceChatId && it.data !is MediaGroupMessage) {
                     messagesListener.send(it)
                 }
             }
+            launch {
+                flowFilter.messageFlow.collectWithErrors(messageUpdatesCollector)
+            }
+            launch {
+                flowFilter.editedMessageFlow.collectWithErrors(messageUpdatesCollector)
+            }
+            launch {
+                flowFilter.channelPostFlow.collectWithErrors(messageUpdatesCollector)
+            }
+            launch {
+                flowFilter.editedChannelPostFlow.collectWithErrors(messageUpdatesCollector)
+            }
 
-            allCallbackQueryListener.subscribe(
-                scope = this
-            ) {
-                (it.data as? MessageDataCallbackQuery) ?.also { query ->
-                    if (query.message.chat.id == config.sourceChatId) {
-                        callbackQueryListener.send(it)
+            launch {
+                flowFilter.callbackQueryFlow.collectWithErrors {
+                    allCallbackQueryListener.offer(it)
+                    (it.data as? MessageDataCallbackQuery) ?.also { query ->
+                        if (query.message.chat.id == config.sourceChatId) {
+                            callbackQueryListener.send(it)
+                        }
                     }
                 }
             }
-            allMediaGroupsListener.subscribe(
-                scope = this
-            ) { mediaGroup ->
+
+            val mediaGroupUpdatesCollector: UpdateReceiver<MediaGroupUpdate> = { mediaGroup ->
+                allMediaGroupsListener.offer(mediaGroup)
                 val mediaGroupChatId = mediaGroup.data.first().chat.id
                 if (mediaGroupChatId == config.sourceChatId) {
                     mediaGroupsListener.send(mediaGroup)
                 }
             }
-
-            val filter = config.createFilter(
-                allMessagesListener,
-                allMessagesListener,
-                allMediaGroupsListener,
-                allCallbackQueryListener
-            )
+            launch {
+                flowFilter.messageMediaGroupFlow.collectWithErrors(mediaGroupUpdatesCollector)
+            }
+            launch {
+                flowFilter.channelPostMediaGroupFlow.collectWithErrors(mediaGroupUpdatesCollector)
+            }
 
             config.subscribe(
-                filter,
+                flowFilter.filter,
                 this
             )
         }