+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/choosers/RateChooser.kt
@@ -3,17 +3,16 @@ package com.github.insanusmokrassar.AutoPostTelegramBot.plugins.choosers
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.models.FinalConfig
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.PluginManager
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.abstractions.Chooser
-import com.github.insanusmokrassar.AutoPostTelegramBot.plugins.rating.RatingPlugin
-import com.github.insanusmokrassar.AutoPostTelegramBot.plugins.rating.database.PostsLikesTable
+import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.abstractions.RatingPlugin
 import com.github.insanusmokrassar.TelegramBotAPI.bot.RequestsExecutor
 
 
 abstract class RateChooser : Chooser {
-    protected var postsLikesTable: PostsLikesTable? = null
+    protected lateinit var ratingPlugin: RatingPlugin
         private set
 
     override suspend fun onInit(executor: RequestsExecutor, baseConfig: FinalConfig, pluginManager: PluginManager) {
         super.onInit(executor, baseConfig, pluginManager)
-        postsLikesTable = (pluginManager.plugins.firstOrNull { it is RatingPlugin } as? RatingPlugin) ?.postsLikesTable
+        ratingPlugin = (pluginManager.plugins.firstOrNull { it is RatingPlugin } as? RatingPlugin) ?: return
     }
 }