+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/scheduler/Scheduler.kt
@@ -5,11 +5,14 @@ import com.github.insanusmokrassar.AutoPostTelegramBot.plugins.publishers.Publis
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.NewDefaultCoroutineScope
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.schedule
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.subscribe
-import kotlinx.coroutines.Job
+import kotlinx.coroutines.*
 import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.launch
 
 private typealias PostTimeToJob = Pair<PostIdPostTime, Job>
+
+private val cancelledException = CancellationException()
+
+private typealias EventLambda = suspend () -> Unit
 
 class Scheduler(
     private val schedulesTable: PostsSchedulesTable,
@@ -18,30 +21,31 @@ class Scheduler(
     private val scope = NewDefaultCoroutineScope(8)
     private var currentPlannedPostTimeAndJob: PostTimeToJob? = null
 
-    private val updateJobChannel = Channel<Unit>(Channel.UNLIMITED)
+    private val updateJobChannel = Channel<EventLambda>(Channel.CONFLATED)
+
+    private val updateLambda: EventLambda = {
+        val replaceBy: PostTimeToJob? = schedulesTable.nearPost() ?.let { nearEvent ->
+            val current = currentPlannedPostTimeAndJob
+            if (current == null || current.first.second.millis != nearEvent.second.millis) {
+                nearEvent to createScheduledJob(nearEvent)
+            } else {
+                current
+            }
+        }
+        if (replaceBy != currentPlannedPostTimeAndJob) {
+            currentPlannedPostTimeAndJob ?.second ?.cancel(cancelledException)
+            currentPlannedPostTimeAndJob = replaceBy
+        }
+    }
 
     private val updateJob: Job = scope.launch {
         for (event in updateJobChannel) {
             try {
-                schedulesTable.nearPost() ?.let { nearEvent ->
-                    val scheduleNew = currentPlannedPostTimeAndJob ?.let { (currentTime, currentJob) ->
-                        if (currentTime.second.millis != nearEvent.second.millis) {
-                            currentJob.cancel()
-                            true
-                        } else {
-                            false
-                        }
-                    } ?: true
-                    if (scheduleNew) {
-                        currentPlannedPostTimeAndJob = nearEvent to createScheduledJob(nearEvent)
-                    }
-                } ?: (currentPlannedPostTimeAndJob ?.second ?.cancel() ?.also {
-                    currentPlannedPostTimeAndJob = null
-                })
+                event()
             } catch (e: Exception) {
                 commonLogger.throwing(
                     Scheduler::class.java.simpleName,
-                    "update job",
+                    "Update job event handling",
                     e
                 )
             }
@@ -50,26 +54,25 @@ class Scheduler(
 
     init {
         schedulesTable.postTimeRegisteredChannel.subscribe {
-            updateJobChannel.send(Unit)
+            updateJobChannel.send(updateLambda)
         }
         schedulesTable.postTimeChangedChannel.subscribe {
-            updateJobChannel.send(Unit)
+            updateJobChannel.send(updateLambda)
         }
         schedulesTable.postTimeRemovedChannel.subscribe {
-            updateJobChannel.send(Unit)
+            updateJobChannel.send(updateLambda)
         }
-        schedulesTable.nearPost() ?.also {
-            updateJobChannel.offer(Unit)
-        }
+        updateJobChannel.offer(updateLambda)
     }
 
     private fun createScheduledJob(by: PostIdPostTime): Job {
         return scope.schedule(
             by.second.millis
         ) {
-            publisher.publishPost(by.first)
-            schedulesTable.unregisterPost(by.first)
-            updateJobChannel.send(Unit)
+            updateJobChannel.send {
+                publisher.publishPost(by.first)
+                schedulesTable.unregisterPost(by.first)
+            }
         }
     }
 }